packages/core/pyproject.toml
[project]
name = "agentshield-core"
version = "1.0.0"
description = "AgentShield Core Engine — AI Agent Runtime Security Platform"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "clickhouse-connect>=0.8",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "cryptography>=44.0",
    "pynacl>=1.5",
    "numpy>=2.0",
    "scikit-learn>=1.5",
    "openai>=1.60",
    "anthropic>=0.42",
    "python-jose>=3.5.0",
    "socksio>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "httpx",  # for TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentshield_core"]
